# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/environments": {
            "get": {
                "description": "返回启用与停用环境，供发布选择和历史记录统一展示；发起发布时仍会校验 enabled。",
                "tags": [
                    "Environment"
                ],
                "summary": "获取完整环境目录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/environment.View"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/v1/system/environments": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "获取系统环境目录",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/environment.View"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "System"
                ],
                "summary": "创建环境",
                "parameters": [
                    {
                        "description": "环境信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/environment.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "$ref": "#/definitions/environment.View"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/system/environments/{code}": {
            "patch": {
                "tags": [
                    "System"
                ],
                "summary": "更新环境",
                "parameters": [
                    {
                        "type": "string",
                        "description": "环境代码",
                        "name": "code",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "待更新字段",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/environment.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        },
                                        "result": {
                                            "$ref": "#/definitions/environment.View"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseTemplate"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "environment.CreateRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "environment.UpdateRequest": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "environment.View": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
...
        "util.ResponseTemplate": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "此处约定：1代表成功，0代表失败",
                    "type": "integer"
                },
                "error": {
                    "description": "如果请求失败，这里一定要给出错误的信息"
                },
                "help": {
                    "description": "显示接口文档地址，便于别人排错",
                    "type": "string"
                },
                "message": {
                    "description": "对请求结果的描述消息，可以为空",
                    "type": "string"
                },
                "result": {
                    "description": "如果请求成功，这里给出成功的结果"
                }
            }
        },
...
    }
}